requirements.txt
# These requirements are for the https://oauth-dropins.appspot.com/ demo app,
# not the library! The library's dependencies are specified in setup.py.
#
# Note that App Engine / Cloud Build's default pip cache is overly aggressive
# with GitHub dependency URLs like these. If I push a new commit to the branch
# in one of these URLs, Cloud Build still uses the old cached version.
# https://cloud.google.com/appengine/docs/standard/python3/specifying-dependencies
#
# Use this to bust that cache: gcloud -q beta app deploy --no-cache ...

git+https://github.com/dvska/gdata-python3.git#egg=gdata

beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
detoxify @ git+https://github.com/unitaryai/detoxify@452932a1662e1d25d4f06db4f0111bbeb132ba40
domain2idna==1.12.0
ecdsa==0.18.0
extras==1.0.0
fixtures==4.0.1
Flask==2.2.2
Flask-Caching==2.0.1
flask-gae-static==1.0
gdata==4.0.1
google-api-core==2.10.2
google-auth==2.6.3
google-cloud-appengine-logging==1.1.6
google-cloud-audit-log==0.2.4
google-cloud-core==2.2.3
google-cloud-datastore==1.15.4
google-cloud-logging==3.3.0
google-cloud-ndb==1.12.0
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
gunicorn==20.1.0
html5lib==1.1
humanize==4.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
Mastodon.py==1.8.0
mf2py==1.1.2
mf2util==0.5.1
mox3==1.1.0
oauthlib==3.2.2
packaging==21.3
pbr==5.11.0
praw==7.6.1
prawcore==2.3.0
proto-plus==1.22.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymemcache==4.0.0
pyparsing==3.0.9
python-tumblpy==1.1.4
pytz==2022.6
PyYAML==6.0
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
testtools==2.5.0
tlslite-ng==0.7.6
transformers==4.24.0
tweepy==4.12.1
ujson==5.4.0
update-checker==0.18.0
urllib3==1.26.13
webapp2==3.0.0b1
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.4.2
Werkzeug==2.2.2
wrapt==1.14.1
